01_homework/requirements.txt

snap-stanford

matplotlib
scipy
numpy
